Extract fares configuration
